+++ Flow/Changes.md
@@ -78,7 +78,7 @@
 - **ModelDeclarations** now support optional properties
 - Added **CachingMap**, **LazyTree** and **LazyInitIterator**
 - Added **ViewGraphNode** as a lazily initialized graph
-- Added generic Range traits: **Span**, **IterableSpan**, **Spanning**, **HasEnds** and so on
+- Added generic Range traits: **Span**, **IterableSpan**, **NumericSpan**, **HasEnds**, and so on
 - Added **TimedTask** trait, which is now supported in **TimedTasks**, also
 - Added new ways to write **LocalTime** values (after importing **TimeExtensions**)
 - Added **SettableOnce**, which is like a **Promise** with **ChangeEvent**s